requirements.txt
feedparser==6.0.11
requests==2.32.3
pandas==2.2.2
supabase==2.7.4
python-dotenv==1.0.1
loguru==0.7.2
spacy==3.7.5
pytest==8.2.2
pytest-cov==5.0.0
httpx==0.27.0
python-dateutil==2.9.0
pytz==2024.1
praw==7.7.1